Refactor magnet mechanism again Major changes:  - mechanisms for fetching metadata in the 'Add torrent' window and fetching after adding are separated  - simplifying asynchronous fetching metadata on the jlibtorrent library side (thank you @aldenml)  - cache stores temp metadata in memory and not physically as a file as before

+++ app/src/main/java/org/proninyaroslav/libretorrent/core/utils/TorrentUtils.java
@@ -61,38 +61,48 @@ public class TorrentUtils
      */
 
     @Nullable
-    public static String copyTorrent(Context context,
-                                     String dataDirId,
-                                     String pathToTorrent) throws Throwable
+    public static String torrentToDataDir(Context context,
+                                           String dataDirId,
+                                           String pathToTorrent) throws Throwable
     {
 
-        if ((pathToTorrent == null ||
-                TextUtils.isEmpty(pathToTorrent)) ||
-                !FileIOUtils.fileExist(pathToTorrent)) {
+        if (pathToTorrent == null || TextUtils.isEmpty(pathToTorrent) || !FileIOUtils.fileExist(pathToTorrent))
             throw new FileNotFoundException();
-        }
 
-        if (torrentDataExists(context, dataDirId)) {
-            String dataDir = findTorrentDataDir(context, dataDirId);
-            /* The same file */
-            String pathToDataTorrent = (dataDir + File.separator + TorrentStorage.Model.DATA_TORRENT_FILE_NAME);
-            if (dataDir != null && pathToDataTorrent.equals(pathToTorrent)) {
-                throw new FileAlreadyExistsException();
+        return torrentToDataDir(context, dataDirId, pathToTorrent, null);
+    }
 
-            } else {
-                removeTorrentDataDir(context, dataDirId);
-            }
-        }
+    @Nullable
+    public static String torrentToDataDir(Context context,
+                                           String dataDirId,
+                                           byte[] bencode) throws Throwable
+    {
+        if (bencode == null)
+            throw new NullPointerException();
 
-        String dataDir = makeTorrentDataDir(context, dataDirId);
+        return torrentToDataDir(context, dataDirId, null, bencode);
+    }
 
-        if (dataDir == null) {
+    private static String torrentToDataDir(Context context, String dataDirId,
+                                           String pathToTorrent, byte[] bencode) throws Throwable
+    {
+        String dataDir;
+        if (torrentDataExists(context, dataDirId))
+            dataDir = findTorrentDataDir(context, dataDirId);
+        else
+            dataDir = makeTorrentDataDir(context, dataDirId);
+        if (dataDir == null)
             throw new IOException("Unable to create dir");
-        }
 
+        /* The same file */
+        if (new File(dataDir, TorrentStorage.Model.DATA_TORRENT_FILE_NAME).exists())
+            throw new FileAlreadyExistsException();
         File torrent = new File(dataDir, TorrentStorage.Model.DATA_TORRENT_FILE_NAME);
-
-        FileUtils.copyFile(new File(pathToTorrent), torrent);
+        /* We are sure that one of them is not null */
+        if (pathToTorrent != null)
+            FileUtils.copyFile(new File(pathToTorrent), torrent);
+        else
+            FileUtils.writeByteArrayToFile(torrent, bencode);
 
         return (torrent.exists() ? torrent.getAbsolutePath() : "");
     }
